Rename module to 'onepay', and prune unused code
