--- app/build.gradle
@@ -1,41 +1,0 @@
-apply plugin: 'com.android.library'
-
-android {
-    flavorDimensions("default")
-    compileSdkVersion 26
-    defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 26
-        versionCode 1
-        versionName "1.0"
-    }
-
-    sourceSets.main {
-        manifest.srcFile 'src/main/AndroidManifest.xml'
-        java.srcDirs = ['src/main/java']
-        resources.srcDirs = ['src/main/res']
-        res.srcDirs = ['src/main/res']
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-
-    productFlavors {
-        ionixdev {}
-        ionixqa {}
-        tbkdev {}
-        tbkqa {}
-        prod {}
-    }
-}
-
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:26.1.0'
-    compile 'com.squareup.retrofit2:retrofit:2.3.0'
-    compile 'com.squareup.retrofit2:converter-jackson:2.3.0'
-}
