--- app/src/tbkdev/java/cl/ionix/tbk_ewallet_sdk_android/http/Constants.java
@@ -1,10 +1,0 @@
-package cl.ionix.tbk_ewallet_sdk_android.http;
-
-/**
- * Created by psepulveda on 06/04/18.
- */
-
-public class Constants {
-    protected static final String BASE_URL = "https://web2desa.test.transbank.cl/ewallet-plugin-api-services/services/";
-    protected static final String GET_TRANSACTION_NUMBER_PATH = "transactionservice/gettransactionnumber";
-}
