+++ onepay/build.gradle
@@ -1,0 +1,30 @@
+apply plugin: 'com.android.library'
+
+android {
+    flavorDimensions("default")
+    compileSdkVersion 26
+    defaultConfig {
+        minSdkVersion 14
+        targetSdkVersion 26
+        versionCode 1
+        versionName "1.0"
+    }
+
+    sourceSets.main {
+        java.srcDirs = ['src/main/java']
+        resources.srcDirs = ['src/main/res']
+        res.srcDirs = ['src/main/res']
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+}
+
+dependencies {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'com.android.support:appcompat-v7:26.1.0'
+}
