+++ onepay/src/main/java/cl/ionix/tbk_ewallet_sdk_android/Error.java
