+++ onepay/src/main/res/values/strings.xml
